Fix compiler warnings
Code was actully well defined in C and gave no warnings, but C++
is a different language and some constructs which are well defined in C
are forbidden in C++. Here I've removed all instances of a nested
anonymous structs/unions/enums, by moving them to the outer scope. This
shouldn't change semantics of the code for both C and C++, and will make
compiler happy.