Add a workaround for manual entry during tech selection
One possibility is to add a separate event: Manual Entry Intent,
and if it is received then nexo-FAST should as for PAN and expiry date.
It looks like Manual Entry maybe needs to be removed from Tech Selection, but
the how to handle eg. MIFARE cards that act as manual entry?